requirements.txt
streamlit>=1.32.0
groq>=0.4.2
faiss-cpu>=1.9.0
sentence-transformers>=2.6.1
newspaper3k==0.2.8
nltk>=3.8.1
vaderSentiment>=3.3.2
scikit-learn>=1.4.1
numpy>=1.26.0
python-dotenv>=1.0.1
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.1.0
plotly>=5.20.0
pandas>=2.2.1
feedparser>=6.0.10
